move qml stuff into new qml folder
