+++ .idea/runConfigurations/Application.xml
@@ -11,7 +11,7 @@
     <option name="IS_MODULE_SDK" value="false" />
     <option name="ADD_CONTENT_ROOTS" value="true" />
     <option name="ADD_SOURCE_ROOTS" value="true" />
-    <option name="SCRIPT_NAME" value="$PROJECT_DIR$/source/ui/main_window.py" />
+    <option name="SCRIPT_NAME" value="C:\Users\cowen\PycharmProjects\hwi_template\source\ui\qml\main_window.py" />
     <option name="PARAMETERS" value="" />
     <option name="SHOW_COMMAND_LINE" value="false" />
     <option name="EMULATE_TERMINAL" value="false" />
