+++ source/sandbox/main_qml_template.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from PySide2.QtGui import QGuiApplication
 from PySide2.QtQml import QQmlApplicationEngine
 from PySide2.QtCore import QUrl
-from source.ui.controllers.simple_daq_controllers import LoginController
+from source.ui.qml.controllers.simple_daq_controllers import LoginController
 from source.application_parameters.application_parameters import LocalFileParameters
 from source.engines.user_workflow_engine import UserWorkflowEngine
