+++ source/ui/qml/controllers/simple_daq_controllers.py
