+++ source/ui/qml/qml_examples/originals/controller.py
