+++ source/ui/qml/qml_examples/originals/main.py
