+++ source/ui/qml/qml_examples/originals/requirements.txt
