+++ source/ui/qml/qml_examples/originals/view.qml
