+++ source/ui/qml/qml_examples/view.qml
