+++ source/ui/qml/qml_views/simple_daq_app_example.qml
